Cleaned up project structure
