--- run.py
@@ -1,3 +1,0 @@
-from Bot.bot import main
-
-main()
